requirements.txt
absl-py==0.9.0
aiodns==1.1.1
aiohttp==3.6.2
appdirs==1.4.3
appnope==0.1.0
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
black==19.10b0
bleach==3.1.3
cachetools==4.0.0
ccxt==1.23.85
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
gym==0.17.1
h5py==2.10.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.5.4
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
TA-Lib==0.4.17
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.0
yarl==1.1.0
zipp==3.1.0
